requirements.txt
llama-cpp-python
Pillow
numpy